Day 6, phase two

+++ src/main/java/io/github/lhug/adventofcode/twentytwentythree/sixth/SingleRace.java
@@ -1,4 +1,4 @@
 package io.github.lhug.adventofcode.twentytwentythree.sixth;
 
-public record SingleRace(int time, int distance) {
+public record SingleRace(long time, long distance) {
 }
